Made calendar preset loader into module
