+++ app/Http/Controllers/Api/PresetController.php
@@ -21,21 +21,6 @@ class PresetController extends Controller
         });
     }
 
-    public function listHtml(Request $request)
-    {
-
-        $presets = '<option>Presets</option>';
-        $presets .= '<option>Random Calendar</option>';
-        $presets .= '<option>Custom JSON</option>';
-        
-        $presets .= Preset::all('id','name')->map(function($preset){
-            return sprintf('<option value="%s">%s</option>', $preset->id, $preset->name);
-        })->implode('');
-
-        return $presets;
-
-    }
-
     public function show(Request $request, $id)
     {
         return Preset::with(['events', 'categories'])->find($id);
