+++ resources/js/app.js
@@ -8,3 +8,4 @@ require('./bootstrap');
 
 window.CalendarClock = require('./clock')
 window.RandomCalendar = require('./random-calendar')
+window.CalendarPresets = require('./calendar-presets')
